Add patient's name to top bar
