+++ app/src/main/res/layout/activity_view_patient_data_packets.xml
@@ -11,7 +11,7 @@
        android:orientation="vertical">
 
        <android.support.v7.widget.Toolbar
-           android:id="@+id/toolbar"
+           android:id="@+id/dataPacketToolbar"
            android:layout_width="match_parent"
            android:layout_height="?attr/actionBarSize"
            android:background="@color/colorPrimary"
